// src/routes/v1/workstationMobile/calendar/read.shiftByDate.api.ts
export interface Request {
  /** @description Ссылка на стейкхолдера */
  stakeholderId: string;

  /** @description Код месяца и года в формате YYYY-MM */
  monthMnemocode: string;

  /** @description Ссылка на торговую точку */
  tradingPointId: string;

  /** @description Текущая дата */
  currentDateFix: string;
}
// ...
export const Errors = {
  // checkUsrSessionMiddleware
  AuthUserMustBeAuthenticated: errors.AuthUserMustBeAuthenticated,
  AuthSessionNotFound: errors.AuthSessionNotFound,
  AuthUsrAccNotFound: errors.AuthUsrAccNotFound,
  AuthUsrAccIsBlocked: errors.AuthUsrAccIsBlocked,
  AuthSessionHasExpired: errors.AuthSessionHasExpired,
  AuthUnableToProlongueSession: errors.AuthUnableToProlongueSession,
  AuthUsrAccMustChangePassword: errors.AuthUsrAccMustChangePassword,
  // checkStakeholderParticipantMiddleware
  AccessCheckStakeholderIsEmpty: errors.AccessCheckStakeholderIsEmpty,
  AccessCheckStakeholderIsBlocked: errors.AccessCheckStakeholderIsBlocked,
  AccessCheckNoStakeholderParticipantRole: errors.AccessCheckNoStakeholderParticipantRole,
  // handler
  GenericLoadEntityProblem: errors.GenericLoadEntityProblem,
  WorkingMonthlyDateNotMatchWithMonthMnemocode: errors.WorkingMonthlyDateNotMatchWithMonthMnemocode,
};